+++ addon/components/flexberry-sitemap.js
@@ -1,0 +1,66 @@
+/**
+  @module ember-flexberry
+*/
+
+import Component from '@ember/component';
+
+/**
+  Component for sitemap render from the object with links.
+
+  @example
+    templates/my-form.hbs
+    ```handlebars
+    {{flexberry-sitemap sitemap=sitemap}}
+    ```
+
+  @class FlexberrySitemapComponent
+  @extends <a href="http://emberjs.com/api/classes/Ember.Component.html">Ember.Component</a>
+*/
+export default Component.extend({
+  /**
+    Object with links description.
+
+    @example
+      ```javascript
+      {
+        nodes: [
+          {
+            link: 'index',
+            caption: 'Home',
+            title: 'Go to homepage!',
+          },
+          {
+            caption: 'Superheroes',
+            children: [
+              {
+                link: 'superman',
+                caption: 'Superman',
+              },
+              {
+                link: 'ironman',
+                caption: 'Ironman',
+              },
+            ],
+          },
+        ],
+      }
+      ```
+
+    @property sitemap
+    @type Object
+  */
+  sitemap: undefined,
+
+  actions: {
+    /**
+      Show or hide menu.
+
+      @method actions.menuToggle
+    */
+    menuToggle() {
+      this.$('.subMenu:first').toggleClass('hidden');
+      this.$('.item-plus:first').toggleClass('hidden');
+      this.$('.item-minus:first').toggleClass('hidden');
+    }
+  },
+});