+++ blueprints/ember-flexberry/index.js
@@ -24,14 +24,6 @@ module.exports = {
       '      paths: [\n' +
       '        \'bower_components/semantic-ui\'\n' +
       '      ]\n' +
-      '    },\n' +
-      '    SemanticUI: {\n' +
-      '      import: {\n' +
-      '        css: false,\n' +
-      '        javascript: true,\n' +
-      '        images: false,\n' +
-      '        fonts: true\n' +
-      '      }\n' +
       '    }\n';
 
     var env1 = '  // Replace this local address to remote when backed will be published.\n' +
@@ -233,13 +225,12 @@ module.exports = {
         { name: 'devicejs', target: '0.2.7' },
         { name: 'blueimp-file-upload', target: '9.11.2' },
         { name: 'flatpickr-calendar', source: 'git://github.com/chmln/flatpickr.git', target: '2.3.4' },
+        { name: 'semantic-ui', target: '2.3.1' },
         { name: 'seiyria-bootstrap-slider', target: '6.0.6' },
         { name: 'jquery-minicolors', target: '2.2.6' },
         { name: 'js-beautify', target: '1.6.4' },
         { name: 'moment', target: '2.22.0' }
       ]);
-    }).then(function() {
-      return _this.addBowerPackageToProject('semantic-ui','git+https://github.com/Flexberry/Semantic-UI.git#fixed-abort');
     }).then(function() {
       return _this.addAddonsToProject({
         packages: [
@@ -252,11 +243,9 @@ module.exports = {
     }).then(function () {
       return _this.addPackagesToProject([
         { name: 'dexie', target: '2.0.2' },
-        { name: 'inflection', target: '1.10.0' },
+        { name: 'inflection', target: '1.12.0' },
         { name: 'node-uuid', target: '1.4.7' },
       ]);
-    }).then(function () {
-      return _this.addPackageToProject('semantic-ui-ember','git+https://github.com/Flexberry/Semantic-UI-Ember.git#version-0.9.3');
     }).then(function () {
       return _this.removePackageFromProject('ember-data');
     }).then(function () {